+++ tests/nestedDynamicZerobuf.cpp
@@ -1,6 +1,6 @@
 
-/* Copyright (c) 2015, Human Brain Project
- *                     Stefan.Eilemann@epfl.ch
+/* Copyright (c) 2015-2016, Human Brain Project
+ *                          Stefan.Eilemann@epfl.ch
  */
 
 #define BOOST_TEST_MODULE nestedDynamicZerobuf
@@ -20,11 +20,14 @@ BOOST_AUTO_TEST_CASE(emptyTestNestedZerobuf)
 BOOST_AUTO_TEST_CASE(copyConstructTestNestedZerobuf)
 {
     test::TestNestedZerobuf temporary;
+    temporary.getDynamic().setName( "Hugo" );
     temporary.getNested().push_back( test::TestNested( 1, 2 ));
     BOOST_CHECK_EQUAL( temporary.getNested().size(), 1 );
+    BOOST_CHECK_EQUAL( temporary.getDynamic().getNameString(), "Hugo" );
 
     const test::TestNestedZerobuf testNestedZerobuf( temporary );
     BOOST_CHECK_EQUAL( testNestedZerobuf.getNested().size(), 1 );
+    BOOST_CHECK_EQUAL( testNestedZerobuf.getDynamic().getNameString(), "Hugo" );
 
     const test::TestNestedZerobuf::Nested& nested=testNestedZerobuf.getNested();
     BOOST_CHECK_EQUAL( nested[0], test::TestNested( 1, 2 ));
@@ -34,19 +37,24 @@ BOOST_AUTO_TEST_CASE(copyConstructTestNestedZerobuf)
 BOOST_AUTO_TEST_CASE(moveConstructTestNestedZerobuf)
 {
     test::TestNestedZerobuf temporary;
+    temporary.getDynamic().setName( "Hugo" );
     temporary.getNested().push_back( test::TestNested( 1, 2 ));
     test::TestNestedZerobuf testNestedZerobuf( std::move( temporary ));
 
     BOOST_CHECK_EQUAL( testNestedZerobuf.getNested().size(), 1 );
     BOOST_CHECK_EQUAL( testNestedZerobuf.getNested()[0],
                        test::TestNested( 1, 2 ));
+    BOOST_CHECK_EQUAL( testNestedZerobuf.getDynamic().getNameString(), "Hugo" );
     BOOST_CHECK_NE( temporary, testNestedZerobuf );
     BOOST_CHECK( temporary.getNested().empty( ));
+    BOOST_CHECK_MESSAGE( temporary.getDynamic().getNameString().empty(),
+                         temporary.getDynamic().getNameString( ));
 
     temporary = std::move( testNestedZerobuf );
     BOOST_CHECK_EQUAL( temporary.getNested().size(), 1 );
     BOOST_CHECK_EQUAL( temporary.getNested()[0],
                        test::TestNested( 1, 2 ));
+    BOOST_CHECK_EQUAL( temporary.getDynamic().getNameString(), "Hugo" );
     BOOST_CHECK_NE( temporary, testNestedZerobuf );
     BOOST_CHECK( testNestedZerobuf.getNested().empty( ));
 
@@ -95,28 +103,65 @@ BOOST_AUTO_TEST_CASE(changeTestNestedZerobuf)
 
 BOOST_AUTO_TEST_CASE(vector)
 {
-    test::TestNestedZerobuf temporary;
-    const test::TestNestedZerobuf& constTemp = temporary;
-    temporary.getNested().push_back( test::TestNested( 1, 2 ));
-    temporary.getNested().push_back( test::TestNested( 3, 4 ));
+    test::TestNestedZerobuf object;
+    const test::TestNestedZerobuf& constObject = object;
+    object.getNested().push_back( test::TestNested( 1, 2 ));
+    object.getNested().push_back( test::TestNested( 3, 4 ));
 
+    BOOST_CHECK_EQUAL( object.getNested().size(), 2 );
+    BOOST_CHECK_EQUAL( object.getNested(), constObject.getNested( ));
+    BOOST_CHECK_THROW( object.getNested()[42], std::runtime_error );
+    BOOST_CHECK_THROW( constObject.getNested()[42], std::runtime_error );
 
-    BOOST_CHECK_EQUAL( temporary.getNested().size(), 2 );
-    BOOST_CHECK_EQUAL( temporary.getNested(), constTemp.getNested( ));
-    BOOST_CHECK_THROW( temporary.getNested()[42], std::runtime_error );
-    BOOST_CHECK_THROW( constTemp.getNested()[42], std::runtime_error );
+    test::TestNestedZerobuf object2;
+    BOOST_CHECK_NE( object.getNested(), object2.getNested( ));
 
-    test::TestNestedZerobuf temporary2;
-    BOOST_CHECK_NE( temporary.getNested(), temporary2.getNested( ));
+    object.getNested().clear();
+    BOOST_CHECK_EQUAL( object.getNested().size(), 0 );
+    BOOST_CHECK( object.getNested().empty( ));
+    BOOST_CHECK_EQUAL( object.getNested(), object2.getNested( ));
+}
 
-    temporary.getNested().clear();
-    BOOST_CHECK_EQUAL( temporary.getNested().size(), 0 );
-    BOOST_CHECK( temporary.getNested().empty( ));
-    BOOST_CHECK_EQUAL( temporary.getNested(), temporary2.getNested( ));
+BOOST_AUTO_TEST_CASE(compact)
+{
+    test::TestNestedZerobuf object;
+    object.getNested().push_back( test::TestNested( 1, 2 ));
+    object.getNested().push_back( test::TestNested( 3, 4 ));
+
+    object.getDynamic().setName( "The quick brown fox" );
+    object.getDynamic().setName( "The quick brown fox jumps" );
+    object.getDynamic().setName( "The quick brown" );
+
+    size_t minSize = test::TestNestedZerobuf::ZEROBUF_STATIC_SIZE() +
+                     2 * test::TestNested::ZEROBUF_STATIC_SIZE() +
+                     test::TestDynamic::ZEROBUF_STATIC_SIZE() +
+                     object.getDynamic().getName().size();
+
+    BOOST_CHECK_LT( minSize, object.getZerobufSize( ));
+    object.compact();
+    object.check();
+    object.getDynamic().check();
+    BOOST_CHECK_EQUAL( minSize, object.getZerobufSize( ));
+
+    minSize = test::TestNestedZerobuf::ZEROBUF_STATIC_SIZE() +
+              test::TestDynamic::ZEROBUF_STATIC_SIZE();
+    object.getDynamic().getName().clear();
+    object.getNested().clear();
+
+    BOOST_CHECK_LT( minSize, object.getZerobufSize( ));
+    object.compact();
+    object.check();
+    object.getDynamic().check();
+    BOOST_CHECK_EQUAL( minSize, object.getZerobufSize( ));
+
 }
 
 const std::string expectedJSON =
     "{\n"
+    "   \"dynamic\" : {\n"
+    "      \"intvalue\" : 7,\n"
+    "      \"name\" : \"Hugo\"\n"
+    "   },\n"
     "   \"nested\" : [\n"
     "      {\n"
     "         \"intvalue\" : 1,\n"
@@ -132,6 +177,7 @@ const std::string expectedJSON =
 BOOST_AUTO_TEST_CASE(testNestedZerobufToGeneric)
 {
     test::TestNestedZerobuf testNestedZerobuf;
+    testNestedZerobuf.getDynamic().setName( "Hugo" );
     testNestedZerobuf.getNested().push_back( test::TestNested( 1, 2 ));
     testNestedZerobuf.getNested().push_back( test::TestNested( 10, 20 ));
     const void* data = testNestedZerobuf.getZerobufData();
@@ -151,6 +197,7 @@ BOOST_AUTO_TEST_CASE(genericToTestNestedZerobuf)
     generic.fromJSON( expectedJSON );
 
     test::TestNestedZerobuf expected;
+    expected.getDynamic().setName( "Hugo" );
     expected.getNested().push_back( test::TestNested( 1, 2 ));
     expected.getNested().push_back( test::TestNested( 10, 20 ));
     const test::TestNestedZerobuf testNestedZerobuf( generic );
@@ -164,6 +211,7 @@ BOOST_AUTO_TEST_CASE(genericToTestNestedZerobuf)
 BOOST_AUTO_TEST_CASE(testNestedZerobufJSON)
 {
     test::TestNestedZerobuf testNestedZerobuf;
+    testNestedZerobuf.getDynamic().setName( "Hugo" );
     testNestedZerobuf.getNested().push_back( test::TestNested( 1, 2 ));
     testNestedZerobuf.getNested().push_back( test::TestNested( 10, 20 ));