+++ tests/serialization.cpp
@@ -1,6 +1,7 @@
 
-/* Copyright (c) 2015, Human Brain Project
- *                     Daniel Nachbaur <danielnachbaur@epfl.ch>
+/* Copyright (c) 2015-2016, Human Brain Project
+ *                          Daniel Nachbaur <daniel.nachbaur@epfl.ch>
+ *                          Stefan.Eilemann@epfl.ch
  */
 
 #define BOOST_TEST_MODULE serialization
@@ -8,7 +9,6 @@
 #include <boost/test/unit_test.hpp>
 
 #include "serialization.h"
-
 
 BOOST_AUTO_TEST_CASE(defaultValues)
 {
@@ -18,6 +18,9 @@ BOOST_AUTO_TEST_CASE(defaultValues)
     BOOST_CHECK_EQUAL( object.getFloatvalue(), 4.2f );
     BOOST_CHECK( !object.getFalseBool( ));
     BOOST_CHECK( object.getTrueBool( ));
+
+    test::TestNestedZerobuf small;
+    BOOST_CHECK_EQUAL( small.getDynamic().getIntvalue(), 7 );
 }
 
 BOOST_AUTO_TEST_CASE(initialized)
@@ -126,4 +129,57 @@ BOOST_AUTO_TEST_CASE(movedVector)
     test::TestNested inner( 1, 2 );
     schema.getNesteddynamic().push_back( inner );
     BOOST_CHECK_EQUAL( inner, schema.getNesteddynamic()[0] );
+    schema.check();
+    schema.getNesteddynamic()[0].check();
+}
+
+BOOST_AUTO_TEST_CASE(compact)
+{
+    test::TestSchema schema = getTestObject();
+    checkTestObject( schema ); // fill _zerobufs in vectors
+
+    size_t oldSize = schema.getZerobufSize();
+    schema.compact(); // compaction under threshold
+    BOOST_CHECK_EQUAL( schema.getZerobufSize(), oldSize );
+    schema.compact( 0.f ); // force compaction
+    BOOST_CHECK_LT( schema.getZerobufSize(), oldSize );
+    checkTestObject( schema ); // fill _zerobufs in vectors
+
+    schema.getNestedMember().setName( "The quick brown fox" );
+    schema.getNestedMember().setName( "The quick brown fox jumps" );
+    schema.getNestedMember().setName( "The quick brown" );
+
+    // Clear allocated memory
+    schema.getIntdynamic().clear();
+    schema.getUintdynamic().clear();
+    schema.getFloatdynamic().clear();
+    schema.getDoubledynamic().clear();
+    schema.getBytedynamic().clear();
+    schema.getShortdynamic().clear();
+    schema.getUbytedynamic().clear();
+    schema.getUshortdynamic().clear();
+    schema.getUlongdynamic().clear();
+    schema.getUint8_tdynamic().clear();
+    schema.getUint16_tdynamic().clear();
+    schema.getUint32_tdynamic().clear();
+    schema.getUint64_tdynamic().clear();
+    schema.getUint128_tdynamic().clear();
+    schema.getInt8_tdynamic().clear();
+    schema.getInt16_tdynamic().clear();
+    schema.getInt32_tdynamic().clear();
+    schema.getInt64_tdynamic().clear();
+    schema.getNesteddynamic().clear();
+    schema.getNestedMember().getName().clear();
+
+    // compact
+    oldSize = schema.getZerobufSize();
+
+    schema.compact();
+    schema.check();
+    schema.getNestedMember().check();
+    BOOST_CHECK_LT( schema.getZerobufSize(), oldSize );
+
+    test::TestSchema copy;
+    copy.fromJSON( schema.toJSON( ));
+    BOOST_CHECK_EQUAL( schema.getZerobufSize(), copy.getZerobufSize( ));
 }