+++ zerobuf/NonMovingAllocator.cpp
@@ -27,15 +27,23 @@ NonMovingAllocator::~NonMovingAllocator()
 
 void NonMovingAllocator::copyBuffer( const void* data, size_t size )
 {
-    _data = (uint8_t*)::realloc( _data, size );
-    _size = size;
+    _resize( size );
     ::memcpy( _data, data, size );
 }
 
-void NonMovingAllocator::_resize( const size_t newSize )
+void NonMovingAllocator::_resize( const size_t size )
 {
-    _data = (uint8_t*)::realloc( _data, newSize );
-    _size = newSize;
+    _data = (uint8_t*)::realloc( _data, size );
+    if( size > _size )
+    {
+        // realloc does not guarantee that additional memory is zero-filled
+#ifdef _WIN32
+        ::memset( _data + _size, 0, size - _size );
+#else
+        ::bzero( _data + _size, size - _size );
+#endif
+    }
+    _size = size;
 }
 
 }