+++ doc/Changelog.md
@@ -1,5 +1,9 @@
 # Changelog {#Changelog}
 
+* [19](https://github.com/HBPVIS/ZeroBuf/pull/19):
+  Allow compaction of existing allocation
+* [19](https://github.com/HBPVIS/ZeroBuf/pull/19):
+  Implemented support for dynamic sub-objects
 * [17](https://github.com/HBPVIS/ZeroBuf/pull/17):
   Implemented support for static sub-objects
 * [17](https://github.com/HBPVIS/ZeroBuf/pull/17):